app/requirements.txt
Flask==1.0.2
Keras==2.2.0
numpy==1.14.5
tensorflow==1.9.0
pillow
